fixup: extract convertion methods
